requirements.txt
torch==2.0.1
torchvision
streamlit
Pillow
os
image
